Auto generate --backend options
